Remove dependency on worker_threads
