+++ src/lib/index.js
@@ -1,4 +1,4 @@
-var convert = require('./workers/worker').convert;
+var convert = require('./workers/conversion').convert;
 
 module.exports = {
     convert: convert
